[UUFV-396] EditOrAddEmployersAndPeriodsReview
